bettingSite/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
-e git+https://github.com/jazzband/django-debug-toolbar.git@783781c8cd396035ac2c2ea2951502f423a9b463#egg=django_debug_toolbar
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
pygame==1.9.6
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-chess==0.30.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.15
webencodings==0.5.1
Willow==1.3
wrapt==1.11.2
XlsxWriter==1.2.8
